Clean up confirm loan by merging the files.
