+++ extension/src/components/ConfirmLoanWizard.tsx
@@ -1,11 +1,30 @@
-import { Alert, AlertDescription, AlertIcon, AlertTitle, Button, Flex, VStack } from "@chakra-ui/react";
+import {
+    Alert,
+    AlertDescription,
+    AlertIcon,
+    AlertTitle,
+    Box,
+    Button,
+    Flex,
+    Heading,
+    Image,
+    Spacer,
+    Text,
+    useInterval,
+    VStack,
+} from "@chakra-ui/react";
 import { Step, Steps, useSteps } from "chakra-ui-steps";
-import * as React from "react";
-import { useState } from "react";
+import Debug from "debug";
+import moment from "moment";
+import React, { useState } from "react";
+import { useAsync } from "react-async";
 import { FiCheck, FiClipboard, FiExternalLink } from "react-icons/all";
-import { confirmLoan, createLoanBackup } from "../background-proxy";
-import { LoanToSign } from "../models";
-import ConfirmLoan from "./ConfirmLoan";
+import { confirmLoan, createLoanBackup, getBlockHeight, rejectLoan, signLoan } from "../background-proxy";
+import { LoanToSign, USDT_TICKER } from "../models";
+import YouSwapItem from "./SwapItem";
+import Usdt from "./tether.svg";
+
+const debug = Debug("confirmloan");
 
 interface ConfirmLoanWizardProps {
     onCancel: () => void;
@@ -19,99 +38,148 @@ const ConfirmLoanWizard = ({ onCancel, onSuccess, loanToSign }: ConfirmLoanWizar
     });
     let [signedTransaction, setSignedTransaction] = useState("");
 
-    const onSigned = (tx: string) => {
-        setSignedTransaction(tx);
-        nextStep();
-    };
+    let { isPending: isSigning, run: sign } = useAsync({
+        deferFn: async () => {
+            let signedTransaction = await signLoan();
+            setSignedTransaction(signedTransaction);
+            nextStep();
+        },
+    });
 
-    const downloadLoanBackup = async () => {
-        const loanBackup = await createLoanBackup(signedTransaction);
-        const file = new Blob([JSON.stringify(loanBackup)], { type: "text/json" });
-        const url = URL.createObjectURL(file);
-        // Note: it would be nicer to open a dialog to download the file. However, we would lose focus
-        // of the window. We don't want that, hence we just open a new tab with the content
-        window.open(url, "_blank");
-        nextStep();
-    };
+    const { isPending: isDownloading, run: downloadLoanBackup } = useAsync({
+        deferFn: async () => {
+            const loanBackup = await createLoanBackup(signedTransaction);
+            const file = new Blob([JSON.stringify(loanBackup)], { type: "text/json" });
+            const url = URL.createObjectURL(file);
+            // Note: it would be nicer to open a dialog to download the file. However, we would lose focus
+            // of the window. We don't want that, hence we just open a new tab with the content
+            window.open(url, "_blank");
+            nextStep();
+        },
+    });
 
-    const onPublish = async () => {
-        await confirmLoan(signedTransaction);
-        onSuccess();
-    };
+    const { isPending: isPublishing, run: publish } = useAsync({
+        deferFn: async () => {
+            await confirmLoan(signedTransaction);
+            onSuccess();
+        },
+    });
 
     return (
         <VStack width="100%">
             <Steps activeStep={activeStep}>
                 <Step label={"Sign"} key={"sign"} icon={FiClipboard}>
                     <Flex py={4}>
-                        <ConfirmLoan
-                            loanToSign={loanToSign}
-                            onCancel={onCancel}
-                            onSigned={onSigned}
-                        />
+                        <form
+                            onSubmit={async e => {
+                                e.preventDefault();
+                                sign();
+                            }}
+                            data-cy="confirm-loan-form"
+                        >
+                            <ConfirmLoan
+                                loanToSign={loanToSign}
+                            />
+                            <Button
+                                variant="secondary"
+                                mr={3}
+                                onClick={async () => {
+                                    await rejectLoan();
+                                    onCancel();
+                                }}
+                            >
+                                Cancel
+                            </Button>
+                            <Button
+                                type="submit"
+                                variant="primary"
+                                isLoading={isSigning}
+                                data-cy="data-cy-sign-loan-button"
+                            >
+                                Sign
+                            </Button>
+                        </form>
                     </Flex>
                 </Step>
                 <Step label={"Backup"} key={"backup"} icon={FiExternalLink}>
                     <Flex py={4}>
                         <VStack>
-                            <Alert
-                                status="warning"
-                                variant="subtle"
-                                flexDirection="column"
-                                alignItems="center"
-                                justifyContent="center"
-                                textAlign="center"
-                                height="250px"
-                            >
-                                <AlertIcon boxSize="40px" mr={0} />
-                                <AlertTitle mt={4} mb={1} fontSize="lg">
-                                    Create a backup!
-                                </AlertTitle>
-                                <AlertDescription>
-                                    Click below to download a backup of the loan details. Together with your seed words
-                                    you can recover your loan details in case your browser storage gets purged. Keep it
-                                    safe! You can restore your backup through the settings.
-                                </AlertDescription>
-                            </Alert>
-                            <Button
-                                onClick={async () => {
-                                    await downloadLoanBackup();
+                            <form
+                                onSubmit={async e => {
+                                    e.preventDefault();
+                                    downloadLoanBackup();
                                 }}
-                                data-cy="data-cy-download-loan-button"
+                                data-cy="confirm-loan-form"
                             >
-                                Download backup
-                            </Button>
+                                <Alert
+                                    status="warning"
+                                    variant="subtle"
+                                    flexDirection="column"
+                                    alignItems="center"
+                                    justifyContent="center"
+                                    textAlign="center"
+                                    height="250px"
+                                >
+                                    <AlertIcon boxSize="40px" mr={0} />
+                                    <AlertTitle mt={4} mb={1} fontSize="lg">
+                                        Create a backup!
+                                    </AlertTitle>
+                                    <AlertDescription>
+                                        Click below to download a backup of the loan details. Together with your seed
+                                        words you can recover your loan details in case your browser storage gets
+                                        purged. Keep it safe! You can restore your backup through the settings.
+                                    </AlertDescription>
+                                </Alert>
+                                <Button
+                                    type="submit"
+                                    variant="primary"
+                                    isLoading={isDownloading}
+                                    data-cy="data-cy-download-loan-button"
+                                >
+                                    Download backup
+                                </Button>
+                            </form>
                         </VStack>
                     </Flex>
                 </Step>
 
                 <Step label={"Confirm"} key={"confirm"} icon={FiCheck}>
                     <Flex py={4}>
-                        <VStack>
-                            <Alert
-                                status="info"
-                                variant="subtle"
-                                flexDirection="column"
-                                alignItems="center"
-                                justifyContent="center"
-                                textAlign="center"
-                                height="200px"
-                            >
-                                <AlertIcon boxSize="40px" mr={0} />
-                                <AlertTitle mt={4} mb={1} fontSize="lg">
-                                    Backup saved?
-                                </AlertTitle>
-                                <AlertDescription>
-                                    The lender will publish the transaction once signed.
-                                </AlertDescription>
-                            </Alert>
-                            <Button
-                                onClick={() => onPublish()}
-                                data-cy="data-cy-confirm-loan-button"
-                            >
-                                Confirm
-                            </Button>
-                        </VStack>
+                        <form
+                            onSubmit={async e => {
+                                e.preventDefault();
+                                publish();
+                            }}
+                            data-cy="confirm-loan-form"
+                        >
+                            <VStack>
+                                <Alert
+                                    status="info"
+                                    variant="subtle"
+                                    flexDirection="column"
+                                    alignItems="center"
+                                    justifyContent="center"
+                                    textAlign="center"
+                                    height="200px"
+                                >
+                                    <AlertIcon boxSize="40px" mr={0} />
+                                    <AlertTitle mt={4} mb={1} fontSize="lg">
+                                        Backup saved?
+                                    </AlertTitle>
+                                    <AlertDescription>
+                                        The lender will publish the transaction once signed.
+                                    </AlertDescription>
+                                </Alert>
+                                <Button
+                                    type="submit"
+                                    variant="primary"
+                                    isLoading={isPublishing}
+                                    data-cy="data-cy-confirm-loan-button"
+                                >
+                                    Confirm
+                                </Button>
+                            </VStack>
+                        </form>
                     </Flex>
                 </Step>
             </Steps>
@@ -120,3 +188,67 @@ const ConfirmLoanWizard = ({ onCancel, onSuccess, loanToSign }: ConfirmLoanWizar
 };
 
 export default ConfirmLoanWizard;
+
+interface ConfirmLoanProps {
+    loanToSign: LoanToSign;
+}
+
+function ConfirmLoan(
+    { loanToSign }: ConfirmLoanProps,
+) {
+    let { details: { collateral, principal, principalRepayment, term } } = loanToSign;
+
+    let { data: blockHeight, reload: reloadBlockHeight } = useAsync({
+        promiseFn: getBlockHeight,
+        onReject: (e) => debug("Failed to fetch block height %s", e),
+    });
+
+    useInterval(() => {
+        reloadBlockHeight();
+    }, 6000); // 1 min
+
+    // format the time nicely into something like : `in 13 hours` or `in 1 month`.
+    // block-height and loan-term are in "blocktime" ^= minutes
+    const deadline = blockHeight && term
+        ? moment().add(Math.abs(blockHeight - term), "minutes").fromNow()
+        : null;
+
+    return (<Box>
+        <Heading>Confirm Loan</Heading>
+        <Box>
+            <YouSwapItem
+                tradeSide={collateral}
+                action={"send"}
+            />
+        </Box>
+        <Box>
+            <YouSwapItem
+                tradeSide={principal}
+                action={"receive"}
+            />
+        </Box>
+        <Box w="100%">
+            <Flex>
+                <Box h="40px" p="1">
+                    <Text>Repayment amount: {principalRepayment}</Text>
+                </Box>
+                <Spacer />
+                <Box w="40px" h="40px">
+                    <Image src={Usdt} h="32px" />
+                </Box>
+                <Box h="40px" justify="right" p="1">
+                    <Text align="center" justify="center">
+                        {USDT_TICKER}
+                    </Text>
+                </Box>
+            </Flex>
+        </Box>
+        <Box w="100%">
+            <Flex>
+                <Box h="40px" p="1">
+                    <Text>Loan term: {term} block-height {deadline ? "(due " + deadline + ")" : ""}</Text>
+                </Box>
+            </Flex>
+        </Box>
+    </Box>);
+}
